# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Invoiça API",
        "description": "API de gestion de devis et factures pour artisans.",
        "contact": {
            "email": "contact@invoica.fr"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Développement"
        },
        {
            "url": "http://my-default-host.com",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/api/v1/documents/{document}/items": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Liste des articles d'un document",
                "operationId": "3ce2b8ed201986bf8a84bd0f4714b4aa",
                "responses": {
                    "200": {
                        "description": "Liste des articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DocumentItem"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Ajouter un article",
                "operationId": "020fc5e7e6f75c0404163a61981be5cf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "designation": {
                                        "type": "string",
                                        "example": "Robinets"
                                    },
                                    "quantity": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 3
                                    },
                                    "unit_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 45.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Article créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentItem"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/documents/{document}/items/{item}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Afficher un article",
                "operationId": "b7b257987286e889ba481efb6fbf285b",
                "responses": {
                    "200": {
                        "description": "Détails de l'article",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentItem"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "Modifier un article",
                "operationId": "af53c5d14a3d98be9f1aa03487cf6af1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "designation": {
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "type": "number"
                                    },
                                    "unit_price": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentItem"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Supprimer un article",
                "operationId": "37102d9dfbe4a3fdeb46ec2e605a4e41",
                "responses": {
                    "200": {
                        "description": "Article supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/company/upload": {
            "post": {
                "tags": [
                    "Fichiers"
                ],
                "summary": "Uploader un fichier (logo, signature, tampon)",
                "operationId": "cc2dd7b3688fc9ed42e8e8084c3b8849",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "logo",
                                            "signature",
                                            "stamp"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Fichier uploadé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "url": {
                                            "type": "string"
                                        },
                                        "path": {
                                            "type": "string"
                                        },
                                        "type": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/company/upload/{type}": {
            "delete": {
                "tags": [
                    "Fichiers"
                ],
                "summary": "Supprimer un fichier uploadé",
                "operationId": "e848a20e6fc84139e10d79217407ce32",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "logo",
                                "signature",
                                "stamp"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/auth/google/callback": {
            "get": {
                "tags": [
                    "Google Auth"
                ],
                "summary": "Callback Google OAuth",
                "operationId": "13e2daf5fc6a3bd2d0bf15b469591861",
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Authentification réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentification Google échouée"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "google_id": {
                        "type": "string",
                        "nullable": true
                    },
                    "avatar": {
                        "type": "string",
                        "nullable": true
                    },
                    "auth_provider": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "company": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Company"
                            }
                        ],
                        "nullable": true,
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Company": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "logo": {
                        "type": "string",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "nullable": true
                    },
                    "address": {
                        "type": "string"
                    },
                    "manager_name": {
                        "type": "string"
                    },
                    "signature": {
                        "type": "string",
                        "nullable": true
                    },
                    "stamp": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "DocumentItem": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "document_id": {
                        "type": "integer"
                    },
                    "designation": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "string"
                    },
                    "unit_price": {
                        "type": "string"
                    },
                    "total_price": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "description": "Entrez votre token Sanctum (Bearer <token>)"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}